+++ conda/_private/shards/subset.py
@@ -64,7 +64,7 @@ import msgpack
 
 import conda.gateways.repodata
 from conda.base.context import context
-from conda.common.compression_zstd import capped_decompress
+from ..zstd import capped_decompress
 
 from . import cache
 from .cache import AnnotatedRawShard
