+++ tests/gateways/test_zstd.py
@@ -14,7 +14,7 @@ from pytest import MonkeyPatch
 
 import conda.gateways.repodata
 from conda.base.context import reset_context
-from conda.common.compression_zstd import zstd
+from conda._private.zstd import zstd
 from conda.core.subdir_data import SubdirData
 from conda.exceptions import CondaHTTPError
 from conda.gateways.connection.session import CondaSession
