+++ tests/shards/conftest.py
@@ -26,7 +26,7 @@ import pytest
 
 from conda._private.shards import cache, shards, subset
 from conda.base.context import context, reset_context
-from conda.common.compression_zstd import zstd
+from conda._private.zstd import zstd
 from conda.models.channel import Channel, all_channel_urls
 
 if TYPE_CHECKING:
