+++ tests/shards/test_shards.py
@@ -41,7 +41,7 @@ from conda._private.shards.shards import (
     shard_mentioned_packages,
 )
 from conda.base.context import context, reset_context
-from conda.common.compression_zstd import zstd
+from conda._private.zstd import zstd
 from conda.core.subdir_data import SubdirData
 from conda.models.channel import Channel
